refactor(features/read): refactor to follow functional core, imperative shell
