+++ src/features/read/mod.rs
@@ -1,5 +1,5 @@
+pub mod domain;
 pub mod handler;
 pub mod logic;
-pub mod domain;
 mod tests;
